requirements.txt
google-api-core==2.19.1  
google-api-python-client==2.139.0 
google-auth==2.32.0  
google-auth-httplib2==0.2.0   
google-auth-oauthlib==1.2.1   
googleapis-common-protos==1.63.2
requests==2.32.3
python-dotenv==1.0.1 